Fixed bountycardback to adjust to darkmode
